+++ build.clj
@@ -45,7 +45,7 @@
    [:scm
     [:url "https://github.com/seancorfield/next-jdbc"]
     [:connection "scm:git:https://github.com/seancorfield/next-jdbc.git"]
-    [:developerConnection "scm:git:ssh://git@github.com/seancorfield/next-jdbc.git"]
+    [:developerConnection "scm:git:ssh:git@github.com:seancorfield/next-jdbc.git"]
     [:tag (str "v" version)]]])
 
 (defn- jar-opts [opts]